Change .java files

+++ src/MainV1.java
@@ -2,7 +2,7 @@ public class MainV1 {
     public static void main(String[] args) {
         BonusMilesService service = new BonusMilesService();
         int price = 10_000;
-        int miles = (int) service.calculate(price);
+        int miles = service.calculate(price);
         System.out.println(miles);
     }
 }
